+++ changelog.md
@@ -2,8 +2,11 @@
 
 ## 0.3.0 - (pre)
 
-* __BREAKING__ Renamed `getTestRunner` to `createUnitTestTask`
 * __BREAKING__ Moved task implementations to new `hop_tasks` library
+
+### hop_tasks
+* __BREAKING__ Renamed `getTestRunner` to `createUnitTestTask`
+* __BREAKING__ Renamed `getDart2jsTask` to `createDart2JsTask`
 
 ## 0.2.1 - 6 Nov 2012 (SDK r14554)